First iteration
Is a working scoreboard, however I've already started working on the array for the buttons by moving the html into javascript innerHTML.  Still need to set up an array for the 1. Home and 2. Guest; as well as another array for the +1, +2, +3 buttons which populate twice.

No stretch goals started yet.